[platform_device_loader] isolate list device related operation
