+++ libparsec/crates/platform_device_loader/src/device/mod.rs
@@ -1,9 +1,11 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
 
 mod archive;
+mod list;
 mod recovery;
 mod save;
 
 pub use archive::*;
+pub use list::*;
 pub use recovery::*;
 pub use save::*;
